refactoring auth to bloc
