+++ lib/domain/services/auth_view_cubit.dart
